Trunkate scroll speed

+++ misc/mouse/mouse.py
@@ -205,7 +205,7 @@ def stop_scroll():
 def scroll_continuous_helper():
     if actions.user.zoom_mouse_idle():
         p = scroll_speed / 100
-        amount = p * scroll_dir * get_scroll_step() / 20
+        amount = int(p * scroll_dir * get_scroll_step() / 20)
         actions.mouse_scroll(by_lines=False, y=amount)
